--- app/src/main/java/alison/fivethingskotlin/adapter/Visibility.kt
@@ -1,8 +1,0 @@
-package alison.fivethingskotlin.adapter
-
-import android.support.annotation.IntDef
-import android.view.View
-
-@IntDef(View.VISIBLE, View.GONE, View.INVISIBLE)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Visibility
