+++ app/src/main/java/alison/fivethingskotlin/fivethings/FiveThingsAdapter.kt
@@ -1,6 +1,5 @@
-package alison.fivethingskotlin.adapter
+package alison.fivethingskotlin.fivethings
 
-import alison.fivethingskotlin.fragment.FiveThingsFragment
 import alison.fivethingskotlin.util.getFullDateFormat
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
@@ -12,7 +11,7 @@ class FiveThingsAdapter(fragmentManager: FragmentManager) : FragmentStatePagerAd
 
     companion object {
         const val PAGER_SIZE = 10000 //user can see data from any day ~13years away from today
-        const val STARTING_DAY = PAGER_SIZE/2
+        const val STARTING_DAY = PAGER_SIZE /2
     }
 
     override fun getCount(): Int {
