+++ app/src/main/java/alison/fivethingskotlin/search/PagedSearchResultAdapter.kt
@@ -1,4 +1,4 @@
-package alison.fivethingskotlin.adapter
+package alison.fivethingskotlin.search
 
 import alison.fivethingskotlin.ContainerActivity
 import alison.fivethingskotlin.model.NetworkState
