+++ app/src/main/res/layout/fragment_calendar.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="viewModel"
-            type="alison.fivethingskotlin.viewmodel.FiveThingsViewModel" />
+            type="alison.fivethingskotlin.fivethings.FiveThingsViewModel" />
     </data>
 
     <FrameLayout
